Authorization policy for deleting threads

+++ app/Http/Controllers/ThreadController.php
@@ -122,7 +122,8 @@ class ThreadController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-        // $thread->replies()->delete();
+        $this->authorize('update', $thread);
+        
         $thread->delete();
 
         if(request()->wantsJson()) {
